requirements.txt
opencv-python
Pillow
numpy
customtkinter
packaging
